+++ src/rateLimiter/fixedWindowCounter/config.go
@@ -1,0 +1,7 @@
+package fixedWindowCounter
+
+// number / time(s: second, m: minute, h:hours)
+// var config = map[string]string
+// 	"anon": "1/s",
+// 	"user": "5/s",
+// }
